requirements.txt
Pillow==11.1.0
pyguetzli==1.0.17
